requirements.txt
black==24.3.0
  click==8.1.7
  mypy-extensions==1.0.0
  packaging==24.0
  pathspec==0.12.1
  platformdirs==4.2.0
  tomli==2.0.1
  typing_extensions==4.10.0
pip-autoremove==0.10.0
  pip==24.0
  setuptools==59.6.0
pipdeptree==2.16.1
pyarr==5.2.0
  overrides==7.7.0
  requests==2.31.0
    certifi==2024.2.2
    charset-normalizer==3.3.2
    idna==3.6
    urllib3==2.2.1
  types-requests==2.31.0.20240311
    urllib3==2.2.1
PyYAML==6.0.1
thefuzz==0.22.1
  rapidfuzz==3.6.2
wheel==0.43.0
